# Vectorize BC4 encoder