--- a/src/encode/bc4.rs
+++ b/src/encode/bc4.rs
@@ -1,6 +1,6 @@
-use glam::Vec4;
+use glam::{UVec4, Vec4};
 
-use crate::{n8, s8, util::clamp_0_1};
+use crate::{n8, s8};
 
 use super::bcn_util::{self, Block4x4};
 
@@ -15,6 +15,45 @@ pub struct Bc4Options {
     pub fast_iter: bool,
 }
 
+struct Block {
+    b: [Vec4; 4],
+}
+impl Block {
+    fn from_raw(block: [f32; 16]) -> Self {
+        let b0 = Vec4::new(block[0], block[1], block[2], block[3]);
+        let b1 = Vec4::new(block[4], block[5], block[6], block[7]);
+        let b2 = Vec4::new(block[8], block[9], block[10], block[11]);
+        let b3 = Vec4::new(block[12], block[13], block[14], block[15]);
+
+        Self {
+            b: [
+                b0.clamp(Vec4::ZERO, Vec4::ONE),
+                b1.clamp(Vec4::ZERO, Vec4::ONE),
+                b2.clamp(Vec4::ZERO, Vec4::ONE),
+                b3.clamp(Vec4::ZERO, Vec4::ONE),
+            ],
+        }
+    }
+    fn min_max(&self) -> (f32, f32) {
+        let [b0, b1, b2, b3] = self.b;
+        let min = b0.min(b1).min(b2).min(b3);
+        let max = b0.max(b1).max(b2).max(b3);
+        (min.min_element(), max.max_element())
+    }
+    fn iter<'a>(&'a self) -> impl Iterator<Item = f32> + 'a {
+        (0..16).map(move |i| self.get_pixel_at(i))
+    }
+}
+impl Block4x4<f32> for &Block {
+    #[inline(always)]
+    fn get_pixel_at(&self, index: usize) -> f32 {
+        debug_assert!(index < 16);
+        let vec_index = (index / 4) % 4;
+        let component_index = index % 4;
+        self.b[vec_index][component_index]
+    }
+}
+
 /// The smallest non-zero value that can be represented in a BC4 block.
 ///
 /// This is also the smallest distance of 2 adjacent representable values.
@@ -22,16 +61,10 @@ const BC4_MIN_VALUE: f32 = 1. / (255. * 7.);
 /// 2 values that are this close will be considered equal.
 const BC4_EPSILON: f32 = 1. / (65536.);
 
-pub(crate) fn compress_bc4_block(mut block: [f32; 16], options: Bc4Options) -> [u8; 8] {
-    // clamp to 0-1
-    block.iter_mut().for_each(|x| *x = clamp_0_1(*x));
+pub(crate) fn compress_bc4_block(block: [f32; 16], options: Bc4Options) -> [u8; 8] {
+    let block = Block::from_raw(block);
 
-    let mut min = block[0];
-    let mut max = block[0];
-    for value in block {
-        min = min.min(value);
-        max = max.max(value);
-    }
+    let (min, max) = block.min_max();
     let diff = max - min;
 
     // reference for testing
@@ -69,9 +102,8 @@ pub(crate) fn compress_bc4_block(mut block: [f32; 16], options: Bc4Options) -> [
 ///
 /// This is intended for testing purposes only. It's EXTREMELY slow. If you do
 /// use it, make sure to enable optimizations.
-fn reference_brute_force(block: [f32; 16]) -> [u8; 8] {
-    let block_min = block.iter().copied().fold(f32::INFINITY, f32::min);
-    let block_max = block.iter().copied().fold(f32::NEG_INFINITY, f32::max);
+fn reference_brute_force(block: Block) -> [u8; 8] {
+    let (block_min, block_max) = block.min_max();
     let min_max = (block_max * 255. + 2.) as u8;
     let max_min = (block_min * 255. - 1.) as u8;
 
@@ -130,7 +162,7 @@ fn single_color(value: f32, options: Bc4Options) -> [u8; 8] {
         let (index_value4, _, error4) = palette4.closest(value);
         let (index_value6, _, error6) = palette6.closest(value);
 
-        if error4.abs() < error6.abs() {
+        if error4 < error6 {
             endpoints4.with_indexes(IndexList::new_all(index_value4))
         } else {
             endpoints6.with_indexes(IndexList::new_all(index_value6))
@@ -181,7 +213,7 @@ fn refine_endpoints(
 }
 
 fn refinement_error_metric<P: Palette>(
-    block: &[f32; 16],
+    block: &Block,
     _options: Bc4Options,
 ) -> impl Copy + Fn((f32, f32)) -> f32 + '_ {
     move |(min, max)| {
@@ -192,7 +224,7 @@ fn refinement_error_metric<P: Palette>(
 }
 
 fn compress_inter6(
-    block: &[f32; 16],
+    block: &Block,
     mut min: f32,
     mut max: f32,
     options: Bc4Options,
@@ -220,10 +252,10 @@ fn compress_inter6(
     (endpoints.with_indexes(indexes), error)
 }
 
-fn compress_inter4(block: &[f32; 16], options: Bc4Options) -> ([u8; 8], f32) {
+fn compress_inter4(block: &Block, options: Bc4Options) -> ([u8; 8], f32) {
     let mut min: f32 = 1.0;
     let mut max: f32 = 0.0;
-    for &value in block {
+    for value in block.iter() {
         if value > BC4_MIN_VALUE {
             min = min.min(value);
         }
@@ -349,7 +381,7 @@ impl EndPoints {
     }
 
     fn new_inter6_unorm(c0: u8, c1: u8) -> Self {
-        assert!(c0 > c1);
+        debug_assert!(c0 > c1);
         let c0_f = n8::f32(c0);
         let c1_f = n8::f32(c1);
         Self { c0, c1, c0_f, c1_f }
@@ -421,23 +453,33 @@ trait Palette {
     where
         Self: Sized;
 
-    /// Returns:
-    /// 0: The index value of the closest color in the palette
-    /// 1: The closest color in the palette
-    /// 2: `pixel - closest`, aka the (signed) error
-    fn closest(&self, pixel: f32) -> (u8, f32, f32);
-
-    /// Returns the square of the error between the pixel and the closest color
-    /// in the palette.
-    ///
-    /// Same as `self.closest(pixel).2.powi(2)`.
-    fn closest_error_sq(&self, pixel: f32) -> f32;
-
     fn from_endpoints(endpoints: &EndPoints) -> Self
     where
         Self: Sized,
     {
         Self::new(endpoints.c0_f, endpoints.c1_f)
+    }
+
+    /// Returns:
+    /// 0: The index value of the closest color in the palette
+    /// 1: The closest color in the palette
+    /// 2: `abs(pixel - closest)`, aka the absolute error
+    fn closest(&self, pixel: f32) -> (u8, f32, f32);
+
+    /// Same as calling `self.closest` for each component of the vector.
+    /// Returns:
+    /// 0: The index values of the closest colors in the palette
+    /// 1: `abs(pixel - closest)`, aka the absolute error
+    fn closest_4(&self, pixels: Vec4) -> (UVec4, Vec4) {
+        let (i0, _, e0) = self.closest(pixels.x);
+        let (i1, _, e1) = self.closest(pixels.y);
+        let (i2, _, e2) = self.closest(pixels.z);
+        let (i3, _, e3) = self.closest(pixels.w);
+
+        (
+            UVec4::new(i0 as u32, i1 as u32, i2 as u32, i3 as u32),
+            Vec4::new(e0, e1, e2, e3),
+        )
     }
 
     /// Returns the index list of the colors in the palette that together
@@ -449,14 +491,16 @@ trait Palette {
     ///
     /// Note that the MSE is **NOT** normalized. In other words, the result is
     /// 16x the actual MSE.
-    fn block_closest(&self, block: &[f32; 16]) -> (IndexList, f32) {
+    fn block_closest(&self, block: &Block) -> (IndexList, f32) {
         let mut total_error = 0.0;
         let mut index_list = IndexList::new_empty();
-        for (pixel_index, pixel) in block.iter().copied().enumerate() {
-            let (index_value, _, error) = self.closest(pixel);
+        for (pixel_index, pixels) in block.b.iter().enumerate() {
+            let (index_value, error) = self.closest_4(*pixels);
 
-            index_list.set(pixel_index, index_value);
-            total_error += error * error;
+            for i in 0..4 {
+                index_list.set(pixel_index * 4 + i, index_value[i] as u8);
+            }
+            total_error += error.dot(error);
         }
 
         (index_list, total_error)
@@ -466,13 +510,7 @@ trait Palette {
     /// the palette.
     ///
     /// Same as `self.block_closest(block).1`.
-    fn block_closest_error_sq(&self, block: &[f32; 16]) -> f32 {
-        block
-            .iter()
-            .copied()
-            .map(|pixel| self.closest_error_sq(pixel))
-            .sum()
-    }
+    fn block_closest_error_sq(&self, block: &Block) -> f32;
 
     fn block_dither(&self, block: impl Block4x4<f32>) -> (IndexList, f32) {
         let mut index_list = IndexList::new_empty();
@@ -491,7 +529,7 @@ trait Palette {
 }
 
 struct Inter6Palette {
-    // c0: f32,
+    c0: f32,
     c1: f32,
     factor1: f32,
     factor2: f32,
@@ -499,6 +537,16 @@ struct Inter6Palette {
 }
 impl Inter6Palette {
     const INDEX_MAP: [u8; 8] = [1, 7, 6, 5, 4, 3, 2, 0];
+
+    const INDEX_MAP_U32: u32 = {
+        let mut map: u32 = 0;
+        let mut i = 0;
+        while i < 8 {
+            map |= (Self::INDEX_MAP[i] as u32) << (i * 4);
+            i += 1;
+        }
+        map
+    };
 }
 impl Palette for Inter6Palette {
     fn new(c0: f32, c1: f32) -> Self {
@@ -509,7 +557,7 @@ impl Palette for Inter6Palette {
         let add1 = 0.5 - c1 * factor1;
 
         Self {
-            // c0,
+            c0,
             c1,
             factor1,
             factor2,
@@ -525,16 +573,48 @@ impl Palette for Inter6Palette {
         let closest = blend7 as f32 * self.factor2 + self.c1;
         let error = pixel - closest;
 
-        (index_value, closest, error)
+        (index_value, closest, error.abs())
     }
-    fn closest_error_sq(&self, pixel: f32) -> f32 {
-        let blend = pixel * self.factor1 + self.add1;
-        let blend7 = ((blend) as u8).min(7) as f32;
 
-        let closest = blend7 * self.factor2 + self.c1;
-        let error = pixel - closest;
+    fn closest_4(&self, pixels: Vec4) -> (UVec4, Vec4) {
+        let blend = pixels * Vec4::splat(self.factor1) + Vec4::splat(self.add1);
+        let blend7 = blend.min(Vec4::splat(7.0)).as_uvec4();
+        let index_value = UVec4::splat(Self::INDEX_MAP_U32) >> (blend7 * 4) & UVec4::splat(0b1111);
+        let closest = blend7.as_vec4() * self.factor2 + self.c1;
+        let error = (pixels - closest).abs();
 
-        error * error
+        (index_value, error)
+    }
+
+    fn block_closest_error_sq(&self, block: &Block) -> f32 {
+        // compute all min errors in parallel
+        let [b0, b1, b2, b3] = block.b;
+
+        // prepare endpoints for interpolation
+        let c0 = Vec4::splat(self.c0);
+        let cd = Vec4::splat(self.c1 - self.c0);
+
+        // start with c0
+        let mut e0 = (c0 - b0).abs();
+        let mut e1 = (c0 - b1).abs();
+        let mut e2 = (c0 - b2).abs();
+        let mut e3 = (c0 - b3).abs();
+
+        // and now the other 7 colors
+        const FACTORS: [f32; 7] = [1. / 7., 2. / 7., 3. / 7., 4. / 7., 5. / 7., 6. / 7., 1.];
+        for f in FACTORS {
+            let c = c0 + cd * f;
+
+            e0 = e0.min((c - b0).abs());
+            e1 = e1.min((c - b1).abs());
+            e2 = e2.min((c - b2).abs());
+            e3 = e3.min((c - b3).abs());
+        }
+
+        // e0-3 now contain the min error for all pixels
+        // so now just square and add
+        let e = e0 * e0 + e1 * e1 + e2 * e2 + e3 * e3;
+        e.x + e.y + e.z + e.w
     }
 }
 
@@ -568,8 +648,8 @@ impl Palette for Inter4Palette {
         // for the rest, check the palette
         #[allow(clippy::needless_range_loop)]
         for i in 0..6 {
-            let error = pixel - self.colors[i];
-            if error.abs() < min_error.abs() {
+            let error = (pixel - self.colors[i]).abs();
+            if error < min_error {
                 min_error = error;
                 index_value = i as u8;
             }
@@ -578,23 +658,63 @@ impl Palette for Inter4Palette {
         (index_value, self.colors[index_value as usize], min_error)
     }
 
-    fn closest_error_sq(&self, pixel: f32) -> f32 {
-        let p0 = Vec4::new(
-            self.colors[0],
-            self.colors[1],
-            self.colors[2],
-            self.colors[3],
-        );
-        let p1 = Vec4::new(
-            self.colors[4],
-            self.colors[5],
-            self.colors[6],
-            self.colors[7],
+    fn closest_4(&self, pixels: Vec4) -> (UVec4, Vec4) {
+        fn to_bvec4(v: glam::BVec4A) -> glam::BVec4 {
+            let bitmask = v.bitmask();
+            glam::BVec4::new(
+                (bitmask & 1) != 0,
+                (bitmask & 2) != 0,
+                (bitmask & 4) != 0,
+                (bitmask & 8) != 0,
+            )
+        }
+        // this handles the case where pixel is 0 or 1
+        let mut min_error = pixels.min(1.0 - pixels);
+        let mut index_value = UVec4::select(
+            to_bvec4(pixels.cmpge(Vec4::splat(0.5))),
+            UVec4::splat(7),
+            UVec4::splat(6),
         );
 
-        let p0_error = (p0 - Vec4::splat(pixel)).abs().min_element();
-        let p1_error = (p1 - Vec4::splat(pixel)).abs().min_element();
-        let error = p0_error.min(p1_error);
-        error * error
+        // for the rest, check the palette
+        #[allow(clippy::needless_range_loop)]
+        for i in 0..6 {
+            let error = (pixels - self.colors[i]).abs();
+            index_value = UVec4::select(
+                to_bvec4(error.cmplt(min_error)),
+                UVec4::splat(i as u32),
+                index_value,
+            );
+            min_error = min_error.min(error);
+        }
+
+        (index_value, min_error)
+    }
+
+    fn block_closest_error_sq(&self, block: &Block) -> f32 {
+        // compute all min errors in parallel
+        let [b0, b1, b2, b3] = block.b;
+
+        // since all pixels are in the range 0-1, we can initialize the min
+        // error with pixel.min(1.0 - pixel) and then do the other 6 colors
+        let mut e0 = b0.min(1.0 - b0);
+        let mut e1 = b1.min(1.0 - b1);
+        let mut e2 = b2.min(1.0 - b2);
+        let mut e3 = b3.min(1.0 - b3);
+
+        // and now the other 6 colors
+        for i in 0..6 {
+            let c = Vec4::splat(self.colors[i]);
+
+            e0 = e0.min((c - b0).abs());
+            e1 = e1.min((c - b1).abs());
+            e2 = e2.min((c - b2).abs());
+            e3 = e3.min((c - b3).abs());
+        }
+
+        // e0-3 now contain the min error for all pixels
+        // so now just square and add
+        let e = e0 * e0 + e1 * e1 + e2 * e2 + e3 * e3;
+        e.x + e.y + e.z + e.w
     }
 }