While working on making the BC1 encoder faster, I noticed that its performance began to approach that of the BC4 encoder. This really shouldn't be the case, since the BC4 encoder only has 25~33% of the pixel data to encode. So I spend some time optimizing the BC4 encoder.

This PR contains no algorithmic improvements. I just tried to use `Vec4` to process 4 pixels at once whereever possible. This was enough to make encoding almost twice as fast.

| Setting     | Old       | New       | Change   |
| ----------- | --------- | --------- | -------- |
| Fast        | 896.51 µs | 552.84 µs | -37.634% |
| Normal      | 2.0166 ms | 1.0504 ms | -46.727% |
| High        | 4.3385 ms | 2.2257 ms | -48.699% |
| High+Dither | 4.6914 ms | 2.6581 ms | -43.341% |

Slight changes in output are most likely due to slightly lower rounding error of some sums that are being calculated and stored in `Vec4` before the components are summed.

---

I found another loop to vectorize, which made it faster again. (I benchmarked on another machine, so the absolute times are different. However, before the last commit, I reproduced the previous speed change within 1%, so the change numbers are comparable.)

| Setting     | Old       | New       | Change   |
| ----------- | --------- | --------- | -------- |
| Fast        | 633.37 µs | 360.40 µs | -43.132% |
| Normal      | 1.3689 ms | 668.24 µs | -50.941% |
| High        | 2.9610 ms | 1.4468 ms | -50.263% |
| High+Dither | 3.2152 ms | 1.8081 ms | -43.559% |